pyproject.toml
[project]
name = "a2a"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "google-api-python-client>=2.186.0",
    "google-auth-httplib2>=0.2.1",
    "google-auth-oauthlib>=1.2.3",
    "langchain-community>=0.4.1",
    "langchain-core>=1.0.2",
    "langchain-groq>=1.0.0",
    "langchain-mcp-adapters>=0.1.12",
    "langchain-openai>=1.0.1",
    "langgraph>=1.0.2",
    "mcp-use>=1.4.0",
    "wikipedia>=1.4.0",
    "requests>=2.32.5",
    "yfinance>=0.2.66",
    "aisuite[anthropic]>=0.1.6",
    "a2a-sdk[http-server]>=0.3.10",
]
